requirements.txt
Flask==2.3.3
google-cloud-datastore>=2.18.0,<2.22.0
google-cloud-storage==2.10.0
PyJWT==2.8.0
cryptography==41.0.4
requests==2.31.0
python-jose[cryptography]==3.3.0
six==1.16.0
Werkzeug==2.3.7
